Re-organize helioviewer page object interfaces
